chore(mockdata): move out of src folder
