+++ docs/components/datetime-picker.md
@@ -360,7 +360,7 @@ Will emit an `object` containing the following values:
 </div>
 
 <script>
-import { TimePeriods, TimeframeKeys } from '@/utils/KDatePickerMockData'
+import { TimePeriods, TimeframeKeys } from 'utils/mockdata/KDatePickerMockData'
 
 const exampleTimeFrames = [
   {
